+++ src/Api/CombatApi.php
@@ -6,6 +6,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
+
 declare(strict_types=1);
 
 namespace Velkuns\GameTextEngine\Api;
@@ -38,6 +39,7 @@ readonly class CombatApi
                 $logs[$turn] = $this->turn($player, $enemy);
 
                 $this->timeResolver->processTurnOnAll([$player, ...$enemies]);
+                $turn++;
             } while ($player->isAlive() && $enemy->isAlive());
 
             if (!$player->isAlive()) {