+++ CHANGELOG.md
@@ -15,4 +15,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove 1
 ```
 
+
+
 ----
+
+## [0.2.0] - 2025-10-30
+[0.2.0]: https://github.com/velkuns/game-text-engine/compare/0.1.0...master
+### Added
+- Add `isEnd` property on node when is node is end of graph
+- Add GameApi::loadFromJsons() to load json directly
+- Add GameApi::read() to handle read next step of story, with validation
+- Add more tests
+### Changed
+- Remove GameApi properties "Api" suffix
+- Update Readme
+
+----
+
+## [0.1.0] - 2025-10-29
+[0.1.0]: https://github.com/velkuns/game-text-engine/compare/0.1.0...master
+### Added
+- Add Element system to manipulate and play
+- Add Graph part to handle graph story
+- Add API part to ease integration in clients